+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to the "clairvoyant" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## 5.2.0 - 2019-??-??
+## 5.2.0 - 2019-12-13
 
 ### Added
 
 - Added `hits:i/n` to the token move menu.
+- Lunatic preview.
+- `clairvoyant.lunaticGoToFile` command.
+- `clairvoyant.enableLunaticPreview` setting.
 
 ### Changed